requirements.txt
bedrock-agentcore-starter-toolkit
boto3

